add ability to set custom caching manager
